requirements.txt
pillow
tkcalendar
reportlab
matplotlib
razorpay
